requirements.txt
Pillow
langchain
openai
stability-sdk
python-pptx